+++ packages/cocoa-discord-utils/postbuild.script.ts
@@ -1,4 +1,5 @@
-import { exec } from "child_process";
+/// <reference types="node" />
+import { exec } from "node:child_process";
 
 for (const file of process.argv.slice(2)) {
     exec(`npx uglifyjs --compress -o ${file} -- ${file}`);