+++ packages/cocoa-discord-utils/src/base/types.ts
@@ -1,3 +1,5 @@
+import { GlobalCommand } from "../slash";
+
 // * https://stackoverflow.com/a/49910890
 export type NonEmptyArray<T> = T[] & { 0: T };
 export type Awaitable<T> = T | PromiseLike<T>;
@@ -6,7 +8,7 @@ export type ResolvesTo<T> = T | (() => Awaitable<T>);
 
 export type BaseCommand = {
     command: { name: string; description?: string };
-    guild_ids?: string[];
+    guild_ids?: string[] | GlobalCommand;
 };
 
 /** commandName **must equal to** it's value .command.name */