+++ packages/cocoa-discord-utils/src/message/index.ts
@@ -1,2 +1,2 @@
-export * from "./Interfaces";
+export * from "./types";
 export * from "./MessageCenter";