+++ packages/cocoa-discord-utils/src/slash/index.ts
@@ -1,3 +1,3 @@
-export * from "./Interfaces";
+export * from "./types";
 export * from "./SlashCenter";
 export * from "./SlashSync";